feat!: modified the layout to multiple pages
