+++ src/routes/about/+page.svelte
@@ -1,0 +1,11 @@
+<script lang="ts">
+</script>
+
+<section class={`flex h-screen w-full flex-col items-center justify-start bg-light-neutral-20`}>
+	<div
+		class={`flex h-3/4 w-full flex-col items-center justify-end bg-light-neutral-10 shadow shadow-light-neutral-40`}
+	>
+		<div class={`h-full w-full max-w-[1040px] border-b border-light-neutral-30`}></div>
+		<div class={`w-full max-w-[1040px] h-40 min-h-40`}></div>
+	</div>
+</section>
